requirements.txt
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==5.1.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
pycparser==2.21
Pygments==2.15.0
requests==2.28.2
rich==13.3.3
tqdm==4.65.0
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.15.5
